Centralize energy rate limit state
