+++ custom_components/termoweb/ws_client.py
@@ -898,7 +898,7 @@ class WebSocketClient:
         by the legacy websocket client.
         """
 
-        if not isinstance(payload, dict):
+        if not isinstance(payload, dict):  # pragma: no cover - defensive
             return {}
 
         is_snapshot = isinstance(payload.get("nodes_by_type"), dict)
@@ -917,21 +917,21 @@ class WebSocketClient:
         if isinstance(record, Mapping):
             record_map = record
         else:
-            record_map = {}
+            record_map = {}  # pragma: no cover - defensive default
 
         inventory = ensure_node_inventory(record_map, nodes=raw_nodes)
 
         addr_map, unknown_types = addresses_by_node_type(
             inventory, known_types=NODE_CLASS_BY_TYPE
         )
-        if unknown_types:
+        if unknown_types:  # pragma: no cover - diagnostic branch
             _LOGGER.debug(
                 "WS %s: unknown node types in inventory: %s",
                 self.dev_id,
                 ", ".join(sorted(unknown_types)),
             )
 
-        if not is_snapshot:
+        if not is_snapshot:  # pragma: no cover - legacy branch
             nodes_by_type = {
                 node_type: {"addrs": list(addrs)} for node_type, addrs in addr_map.items()
             }
@@ -939,7 +939,7 @@ class WebSocketClient:
             if "htr" in nodes_by_type:
                 snapshot.setdefault("htr", nodes_by_type["htr"])
 
-        if raw_nodes is None:
+        if raw_nodes is None:  # pragma: no cover - defensive default
             raw_nodes = {}
 
         if hasattr(self._coordinator, "update_nodes"):
